requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
black==25.11.0
celery==5.6.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.0
django-environ==0.12.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
IMAPClient==3.0.1
kombu==5.6.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytokens==0.3.0
redis==7.1.0
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14
